Use new `commutative-semigroups` package
Long story short, we just made the package `commutative-semigroups` to factor some
code out of `groups` so we could use `groups` in `patch`. We didn't just
use `commutative-semigroups` because the maintainers of `groups` thought
it would be good to keep `groups` fairly minimal.

It would be nice to keep on deduplicating things beyond `groups` and
`patch`, so we have
https://github.com/obsidiansystems/commutative-semigroups/issues/1 to
track this. Hence this PR.