Inline definition of `cancellativeGCDInstances`.
This list only had a single member: `()`.

Since the `GCDMonoid` class no longer states any kind of distributivity
laws (which were conditional on types also being instances of
`Cancellative`), it's no longer necessary to maintain a separate list of
types that are instances of both `GCDMonoid` and `Cancellative`.

Nevertheless, in order to not lose any existing test coverage, we inline
this list in all the places that it's used.