Simplify existing instance definitions for `Const`.
This commit uses `GeneralizedNewtypeDeriving` to simplify existing class
instance definitions for `Const`.

The approach used is similar to that used in `base`, which uses
`GeneralizedNewtypeDeriving` for `Semigroup` and `Monoid` (among others):

```hs
newtype Const a b = Const { getConst :: a }
    deriving ( ...
             , Semigroup  -- ^ @since base-4.9.0.0
             , Monoid     -- ^ @since base-4.9.0.0
             , ...
             )
```

In practice, this mechanism generates method definitions that use
zero-cost coercions (see `Data.Coerce`).